# doc/swagger.yaml
definitions:
...
  crypto.PublicKeyJWK:
    properties:
      alg:
        type: string
      crv:
        type: string
      e:
        type: string
      key_ops:
        type: string
      kid:
        type: string
      kty:
        type: string
      "n":
        type: string
      use:
        type: string
      x:
        type: string
      "y":
        type: string
    required:
    - kty
    type: object
  did.Document:
    properties:
      '@context': {}
      alsoKnownAs:
        type: string
      assertionMethod:
        items: {}
        type: array
      authentication:
        items: {}
        type: array
      capabilityDelegation:
        items: {}
        type: array
      capabilityInvocation:
        items: {}
        type: array
      controller:
        type: string
      id:
        description: |-
          As per https://www.w3.org/TR/did-core/#did-subject intermediate representations of DID Documents do not
          require an ID property. The provided test vectors demonstrate IRs. As such, the property is optional.
        type: string
      keyAgreement:
        items: {}
        type: array
      service:
        items:
          $ref: '#/definitions/did.Service'
        type: array
      verificationMethod:
        items:
          $ref: '#/definitions/did.VerificationMethod'
        type: array
    type: object
  did.DocumentMetadata:
    properties:
      canonicalId:
        type: string
      created:
        type: string
      deactivated:
        type: boolean
      equivalentId:
        type: string
      nextUpdate:
        type: string
      nextVersionId:
        type: string
      updated:
        type: string
      versionId:
        type: string
    type: object
  did.ResolutionError:
    properties:
      code:
        type: string
      invalidDid:
        type: boolean
      notFound:
        type: boolean
      representationNotSupported:
        type: boolean
    type: object
  did.ResolutionMetadata:
    properties:
      contentType:
        type: string
      error:
        $ref: '#/definitions/did.ResolutionError'
    type: object
  did.Service:
    type: object
  did.VerificationMethod:
    properties:
      blockchainAccountId:
        description: for PKH DIDs - https://github.com/w3c-ccg/did-pkh/blob/90b28ad3c18d63822a8aab3c752302aa64fc9382/did-pkh-method-draft.md
        type: string
      controller:
        type: string
      id:
        type: string
      publicKeyBase58:
        type: string
      publicKeyJwk:
        $ref: '#/definitions/crypto.PublicKeyJWK'
        description: must conform to https://datatracker.ietf.org/doc/html/rfc7517
      publicKeyMultibase:
        description: https://datatracker.ietf.org/doc/html/draft-multiformats-multibase-03
        type: string
      type:
        type: string
    required:
    - controller
    - id
    - type
    type: object
...
  github.com_tbd54566975_ssi-service_pkg_server_router.CreateDIDByMethodRequest:
    properties:
      keyType:
        description: |-
          Identifies the cryptographic algorithm family to use when generating this key.
          One of the following: "Ed25519", "X25519", "secp256k1", "P-224","P-256","P-384", "P-521", "RSA"
        type: string
      options:
        description: Options for creating the DID. Implementation dependent on the
          method.
    required:
    - keyType
    type: object
  github.com_tbd54566975_ssi-service_pkg_server_router.CreateDIDByMethodResponse:
    properties:
      did:
        $ref: '#/definitions/did.Document'
      keyType:
        description: |-
          Identifies the cryptographic algorithm family to use when generating this key.
          One of the following: "Ed25519", "X25519", "secp256k1", "P-224","P-256","P-384", "P-521", "RSA"
        type: string
      privateKeyBase58:
        type: string
    type: object
...
  github.com_tbd54566975_ssi-service_pkg_server_router.GetDIDByMethodResponse:
    properties:
      did:
        $ref: '#/definitions/did.Document'
    type: object
...
  github.com_tbd54566975_ssi-service_pkg_server_router.GetDIDsByMethodResponse:
    properties:
      dids:
        items:
          $ref: '#/definitions/did.Document'
        type: array
    type: object
  github.com_tbd54566975_ssi-service_pkg_server_router.GetDIDsRequest:
    properties:
      filter:
        description: |-
          A standard filter expression conforming to https://google.aip.dev/160.
          Not implemented yet.
        type: string
    type: object
...
  github.com_tbd54566975_ssi-service_pkg_server_router.ResolveDIDResponse:
    properties:
      didDocument:
        $ref: '#/definitions/did.Document'
      didDocumentMetadata:
        $ref: '#/definitions/did.DocumentMetadata'
      didResolutionMetadata:
        $ref: '#/definitions/did.ResolutionMetadata'
    type: object
...
host: localhost:8080
info:
  contact:
    email: tbd-developer@squareup.com
    name: TBD
    url: https://github.com/TBD54566975/ssi-service/issues
  description: https://github.com/TBD54566975/ssi-service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: SSI Service API
  version: "0.1"
paths:
...
  /v1/dids:
    get:
      consumes:
      - application/json
      description: Get DIDs by method
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.GetDIDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.GetDIDsByMethodResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get DIDs
      tags:
      - DecentralizedIdentityAPI
  /v1/dids/{method}:
    put:
      consumes:
      - application/json
      description: |-
        Creates a DID document with the given method. The document created is stored internally and can be
        retrieved using the GetOperation. Method dependent registration (for example, DID web registration)
        is left up to the clients of this API.
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.CreateDIDByMethodRequest'
      - description: Method
        in: path
        name: method
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.CreateDIDByMethodResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create DID Document
      tags:
      - DecentralizedIdentityAPI
  /v1/dids/{method}/{id}:
    delete:
      consumes:
      - application/json
      description: |-
        When this is called with the correct did method and id it will flip the softDelete flag to true for the db entry.
        A user can still get the did if they know the DID ID, and the did keys will still exist, but this did will not show up in the GetDIDsByMethod call
        This facilitates a clean SSI-Service Admin UI but not leave any hanging VCs with inaccessible hanging DIDs.
        Soft Deletes DID by method
      parameters:
      - description: Method
        in: path
        name: method
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Soft Delete DID
      tags:
      - DecentralizedIdentityAPI
    get:
      consumes:
      - application/json
      description: Get DID by method
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.CreateDIDByMethodRequest'
      - description: Method
        in: path
        name: method
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.GetDIDByMethodResponse'
        "400":
          description: Bad request
          schema:
            type: string
      summary: Get DID
      tags:
      - DecentralizedIdentityAPI
  /v1/dids/resolver/{id}:
    get:
      consumes:
      - application/json
      description: Resolve a DID that may not be stored in this service
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.ResolveDIDResponse'
        "400":
          description: Bad request
          schema:
            type: string
      summary: Resolve a DID
      tags:
      - DecentralizedIdentityAPI
...
swagger: "2.0"